refactor(12): change file name from keys.js to index.js
