+++ lib/utils.js
@@ -1,7 +1,7 @@
 import bcrypt from 'bcrypt';
 import jwt from 'jsonwebtoken';
 
-import { jwtSecret, jwtExpiresIn } from '../config/keys.js';
+import { jwtSecret, jwtExpiresIn } from '../config/index.js';
 
 export const hashPassword = async password => {
   if (typeof password !== 'string' || password.length === 0) {
